+++ packages/localize/src/tests/xlb-init.test.ts
@@ -1,9 +1,9 @@
 /**
  * @license
- * Copyright 2020 Google LLC
+ * Copyright 2021 Google LLC
  * SPDX-License-Identifier: BSD-3-Clause
  */
 
 import {e2eGoldensTest} from './e2e-goldens-test';
 
-e2eGoldensTest('transform', ['--config=lit-localize.json', 'build']);
+e2eGoldensTest('xlb-init', ['--config=lit-localize.json', 'extract']);