+++ packages/localize/testdata/xliff-init/goldens/tsout/locales/es-419.ts
@@ -1,9 +1,0 @@
-// Do not modify this file by hand!
-// Re-generate this file by running lit-localize
-
-/* eslint-disable no-irregular-whitespace */
-/* eslint-disable @typescript-eslint/no-explicit-any */
-
-export const templates = {
-  string: `Hello World!`,
-};